Split save_types.rs (885 lines) into submodule directory
Break the monolithic save_types.rs into five focused files under
save_types/ to improve navigability and reduce merge conflicts:

- version.rs (39 lines): version constants
- core_types.rs (259 lines): grid, road, clock, budget, citizen structs
- policy_types.rs (180 lines): V2 policy, weather, budget, timer structs
- infrastructure_types.rs (314 lines): environment & infrastructure structs
- save_data.rs (114 lines): top-level SaveData struct
- mod.rs: re-exports all pub items for a flat namespace

All callers continue to use `crate::save_types::*` unchanged.

Part of #1389
